\name{prettytree}
\alias{prettytree}
\title{
  Print a tree.
}
\description{
  Produces textual output representing a tree.
}
\usage{
prettytree(x, inames = NULL, ilevels = NULL)
}
\arguments{
  \item{x}{
      a recursive list representing a tree.
}
  \item{inames}{
      optional variable names.
}
  \item{ilevels}{
      an optional list of levels for factors.
}
}
\details{
This function is normally not called by users 
but needed in some reverse dependencies of party.
}
\keyword{tree}
